+++ package.json
@@ -35,7 +35,7 @@
   ],
   "license": "ISC",
   "scripts": {
-    "test": "mocha test",
+    "test": "mocha test/index",
     "cover": "istanbul cover _mocha -- test"
   },
   "dependencies": {
